SW-18: fixing comment

+++ schema-wizard-core/src/main/java/org/schemawizard/core/property/service/PropertyParser.java
@@ -5,5 +5,5 @@ public interface PropertyParser {
 
     boolean parseBooleanValue(String value);
 
-    Integer parseLongValue(String value);
+    Integer parseIntegerValue(String value);
 }
